% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loss_df.R
\name{check_loss_df}
\alias{check_loss_df}
\title{check that arguments and data columns to loss_df are appropriate}
\usage{
check_loss_df(ldf)
}
\arguments{
\item{ldf}{a loss_df to check}
}
\description{
check that arguments and data columns to loss_df are appropriate
}

