% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cum2incr}
\alias{cum2incr}
\title{Transforms cumulative loss payments into incrmental}
\usage{
cum2incr(ldf, value)
}
\arguments{
\item{ldf}{object of class loss_df}

\item{col}{column to transform from cumulative to incremental}
}
\description{
Transforms cumulative loss payments into incrmental
}
\examples{
increamental(recover_ldf, value = "paid")
}

