% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dev_wtd}
\alias{dev_wtd}
\title{returns the weighted average developemt factors}
\usage{
dev_wtd(triangle, increasing = TRUE)
}
\arguments{
\item{triangle}{an object of class triangle defined by `ChainLadder` package}

\item{increasing}{logical just the order of the vector}
}
\description{
returns the weighted average developemt factors
}
\examples{
tri <- as.triangle(recovery_ldf, origin = "origin", dev = "dev", value = "paid_loss_only")
dev_wtd(tri)
}

