% Generated by roxygen2 (4.0.1): do not edit by hand
\name{exhibit.ata}
\alias{exhibit.ata}
\title{Return a triangle for age to age development factors}
\usage{
\method{exhibit}{ata}(ata, selection = NULL)
}
\arguments{
\item{ata}{object of class ata generated from \code{ChainLadder} package}

\item{selection}{optional selected development factors}
}
\description{
Return a triangle for age to age development factors
}
\keyword{internal}

