% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exhibit.glmReserve}
\alias{exhibit.glmReserve}
\title{Returns glmReserve summary information in a data frame}
\usage{
\method{exhibit}{glmReserve}(glmReserve)
}
\arguments{
\item{glmReserve}{object of class glmReserve}
}
\description{
Returns glmReserve summary information in a data frame
}
\keyword{internal}

