% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exhibit.ratio}
\alias{exhibit.ratio}
\title{Returns a ratio triangle for use in reports}
\usage{
\method{exhibit}{ratio}(tri)
}
\arguments{
\item{tri}{object of class ratio and triangle generated from \code{ChainLadder} package}
}
\description{
Returns a ratio triangle for use in reports
}
\keyword{internal}

