% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exhibit}
\alias{exhibit}
\title{Generic similar to summary}
\usage{
exhibit(data, ...)
}
\arguments{
\item{data}{object to turn into exhibit}

\item{...}{additional arguments}
}
\description{
\code{exhibit} defines certain default ways to arrange
objects returned by the package \code{ChainLadder} package
and \code{loss_db} packages for presentation in a report.
A typical result from executing the \code{exhibit} function
is a data frame in which each row represents an origin
period and each column represents a some aggregate claim
loss amount.
}

