% Generated by roxygen2 (4.0.1): do not edit by hand
\name{exhibit.triangle}
\alias{exhibit.triangle}
\title{exhibit generic method for class triangle from ChainLadder package}
\usage{
\method{exhibit}{triangle}(tri)
}
\description{
exhibit generic method for class triangle from ChainLadder package
}

