\docType{data}
\name{general}
\alias{general}
\title{Data descriptive of the workers' compensation claims data in \code{\link{losses}}}
\description{
a data frame detailing premium, payroll, excess reinsurance coverage
and applicable deductibles for the \code{\link{losses}} and \code{\link{losses_loss_df}}
data sets
}

