% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_calendar}
\alias{get_calendar}
\title{return calandar year for each row}
\usage{
get_calendar(df, year = NULL)
}
\arguments{
\item{df}{a loss_df data frame}
}
\description{
useful for finding the evaluation date if one is not supplied
}

