% Generated by roxygen2 (4.0.2): do not edit by hand
\name{incremental}
\alias{incremental}
\title{Transforms cumulative loss payments into incremental}
\usage{
incremental(ldf, dollar)
}
\arguments{
\item{ldf}{object of class loss_df}

\item{dollar}{column to transform from cumulative to incremental. Can be of any of
the allowable `detail` types for class `loss_df`}
}
\description{
Transforms cumulative loss payments into incremental
}
\examples{
incremental(ldf_data, dollar = "paid")
}

