% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sum_dollar.R
\name{incurred_recovery}
\alias{incurred_recovery}
\title{returns a vector for total incurred recovery}
\usage{
incurred_recovery(ldf)
}
\arguments{
\item{ldf}{a data frame of class loss_df}
}
\description{
returns a vector for total incurred recovery
}
\examples{
incurred_recovery(recovery_ldf)
}

