% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{ldf_data}
\alias{ldf_data}
\title{example workers' compensation loss_df}
\format{A \code{\link{loss_df}} object}
\usage{
ldf_data
}
\description{
This is an example \code{\link{loss_df}} S3 class object created from
the \code{losses} data frame.  It was created using the \code{\link{loss_df}} function.
}

