% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loss_df}
\alias{loss_df}
\title{loss_df S3 class constructor}
\usage{
loss_df(ldf, origin, dev, id = NULL, paid = NULL, incurred = NULL,
  paid_recovery = NULL, incurred_recovery = NULL, desc = NULL)
}
\arguments{
\item{ldf}{a data frame containing columns to be passed to \code{loss_df}}

\item{origin}{policy period in which the claim/loss is associated}

\item{dev}{development stage of claim/loss}

\item{id}{optional identification key for loss (e.g. a claim number)}

\item{paid}{vector listing names for all paid loss columns}

\item{incurred}{vector listing names for all incurred loss columns}

\item{paid_recovery}{vector listing names for all paid recovery columns}

\item{incurred_recovery}{vector listing names for all incurred recovery columns}

\item{desc}{vector of all additional columns relevant to the analysis. Common examples are
  "open claims", and "reported claims" (when claims are evaluated on an occurrence basis)}
}
\description{
\code{loss_df} is the primary S3 data class for the \code{lossdb}
package.  A \code{loss_df} is a data frame for insurance loss data.  Each
row/observation represents losses at a unique "origin", "dev", and optional
"id" (i.e \code{loss_df} takes loss data organized by claim/occurrence "id",
"origin", and "dev" or just by "origin" and "dev").  Each column contains
certain variables of interest to the actuarial analysis.
}
\details{
\code{loss_df} is designed to assign descriptive and consistent names
to loss data to allow for quick execution of common actuarial analysis tasks.
}
\examples{
# create loss_df object from `occurences` dataset
mydf <- loss_df(occurences,  id = "claim_number",
                             origin = "origin",
                             dev = "dev",
                             paid = c("paid_loss_only", "paid_expense"),
                             incurred = c("incurred_loss_only", "incurred_expense"),
                             paid_recovery = c("sal_sub", "paid_excess250"),
                             incurred_recovery = c("incurred_excess250", "sal_sub_incurred"),
                             desc = "claim_cts"
                 )
}

