% Generated by roxygen2 (4.0.0): do not edit by hand
\name{losses}
\alias{losses}
\title{example workers' compensation \code{\link{loss_df}}}
\format{A \code{\link{loss_df}} object  with 8,054 observations on 9 variables.}
\usage{
losses_loss_df

losses
}
\description{
This is a cleaned version of the \code{losses} data frame.  It was created
using the \code{\link{loss_df}} function.

This claims data is in a 'raw' state; it is constructed
of 47 columns, most of which are not relevant to a reserve analysis.  The
\code{losses} data set is meant to resemble the combination of 5 typical data sets
an actuary would receive from a small workers' compensation insurance provider.  The \code{losses}
claims data is evaluated at 2009-06-30, 2010-06-30, 2011-06-30, 2012-06-30, and
2013-06-30.
}

