\name{losses_ldf}
\alias{losses_ldf}
\title{example workers' compensation \code{\link{loss_df}}}
\format{A \code{\link{loss_df}} object  with 8,054 observations on 9 variables.}
\usage{
losses_ldf
}
\description{
This is a cleaned version of the \code{losses} data frame.  It was created
using the \code{\link{loss_df}} function.
}

