% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_ldf}
\alias{merge_ldf}
\title{merge wrapper for data frame using 'type' attribute}
\usage{
merge_ldf(df, calendar1, calendar2, by, exclude)
}
\arguments{
\item{df}{}

\item{by}{vector of column names to merge by}

\item{columns}{to be excluded from the merge}
}
\description{
merge wrapper for data frame using 'type' attribute
}

