% Generated by roxygen2 (4.0.2): do not edit by hand
\name{merge_loss_df}
\alias{merge_loss_df}
\title{merge wrapper for data frame using 'type' attribute}
\usage{
merge_loss_df(df, calendar1, calendar2, by, exclude)
}
\arguments{
\item{df}{}

\item{calendar1}{}

\item{calendar2}{}

\item{by}{vector of column names to merge by}

\item{columns}{to be excluded from the merge}
}
\description{
merge wrapper for data frame using 'type' attribute
}

