% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occurrences}
\alias{occurrences}
\title{The \code{\link{losses}} data frame transformed for use in the \code{\link{loss_df}} function}
\description{
Columns have been transformed from dates to origin years and
development years.  Also losses have been split up as desired for the examples.
}

