% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sum_dollar.R
\name{paid}
\alias{paid}
\title{returns a vector for total gross paid}
\usage{
paid(ldf)
}
\arguments{
\item{ldf}{a data frame of class loss_df}
}
\description{
returns a vector for total gross paid
}
\examples{
paid(recovery_ldf)
}

