% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sum_dollar.R
\name{paid_recovery}
\alias{paid_recovery}
\title{returns a vector for total paid recovery}
\usage{
paid_recovery(ldf)
}
\arguments{
\item{ldf}{a data frame of class loss_df}
}
\description{
returns a vector for total paid recovery
}
\examples{
paid_recovery(recovery_ldf)
}

