% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loss_df.R
\name{plot.loss_df}
\alias{plot.loss_df}
\title{plot method for \code{loss_df}}
\usage{
\method{plot}{loss_df}(ldf, calendar = NULL)
}
\arguments{
\item{df}{loss_df S3 object}
}
\description{
barchart of loss amounts by origin year
}
\examples{
# plot paid and case at most recent `calendar`
plot(ldf_data)

# plot paid and case at selected `calendar`
plot(ldf_data, calendar = 2012)
}
\keyword{internal}

