% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recovery_ldf}
\alias{recovery_ldf}
\title{example workers' compensation loss_df}
\format{A \code{\link{loss_df}} object  with 8,054 observations on 12 variables.}
\usage{
losses_ldf
}
\description{
This is a cleaned version of the \code{losses} data frame.  It was created
using the \code{\link{loss_df}} function.
}

