% Generated by roxygen2 (4.0.2): do not edit by hand
\name{retain}
\alias{retain}
\title{Return retained amounts by claim}
\usage{
retain(ldf, value, recovery = NULL)
}
\arguments{
\item{ldf}{data frame of class loss_df}

\item{value}{}

\item{recovery}{}
}
\description{
Return retained amounts by claim
}
\examples{
head(retain(recovery_ldf, value = "paid", recovery = c("paid_excess250", "sal_sub")))

head(retain(recovery_ldf, value = "paid", recovery = "paid_recovery"))
}

