% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sum_by_type}
\alias{sum_by_type}
\title{returns the loss df with each type attribute summed up}
\usage{
sum_by_type(ldf, sum_types)
}
\arguments{
\item{ldf}{a data frame of class loss_df}

\item{sum_types}{the types to be summed up.  Possible `sum_types` are `paid`, `incurred`, `paid_recovery`, and `incurred_recovery`.}
}
\description{
returns the loss df with each type attribute summed up
}
\examples{
by_type(ldf = recovery_ldf, sum_types = "paid")
by_type(ldf = recovery_ldf, sum_types = c("paid", "incurred")
}

