% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sum_type}
\alias{sum_type}
\title{returns the sum of the selected loss_df type}
\usage{
sum_type(df, type)
}
\arguments{
\item{type}{the type to be summed up. `type` options are `paid`, `incurred`, `paid_recovery`, and `incurred_recovery`.}

\item{ldf}{a data frame of class loss_df}
}
\description{
returns the sum of the selected loss_df type
}
\examples{
sum_type(ldf = recovery_ldf, sum_types = "paid")
sum_type(ldf = recovery_ldf, sum_types = c("paid", "incurred")
}

