% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{test_df_incomplete}
\alias{test_df_incomplete}
\title{Minimal loss_df with missing historic data}
\format{A \code{\link{loss_df}} object}
\usage{
test_df_incomplete
}
\description{
A small \code{\link{loss_df}} with missing historic data
used to test and explain functions.
}

