% Generated by roxygen2 (4.0.2): do not edit by hand
\name{type_totals}
\alias{type_totals}
\title{Totals the type attribues `paid`, `incurred`, `paid_recovery`, `incurred_recovery`}
\usage{
type_totals(ldf)
}
\arguments{
\item{ldf}{object of type loss_df}
}
\description{
Totals the type attribues `paid`, `incurred`, `paid_recovery`, `incurred_recovery`
}
\examples{
full_ldf(ldf_data)
}

