% Generated by roxygen2 (4.0.0): do not edit by hand
\name{check_loss_df}
\alias{check_loss_df}
\title{check that arguments and data columns to loss_df are appropriate}
\usage{
check_loss_df(df)
}
\arguments{
\item{df}{a loss_df to check}
}
\description{
check that arguments and data columns to loss_df are appropriate
}

