% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cum2incr}
\alias{cum2incr}
\title{Transforms cumulative loss payments into incrmental}
\usage{
cum2incr(ldf, detail)
}
\arguments{
\item{ldf}{object of class loss_df}

\item{detail}{column to transform from cumulative to incremental. Can be of any of
the allowable `detail` types for class `loss_df`}
}
\description{
Transforms cumulative loss payments into incrmental
}
\examples{
incremental(recover_ldf, detail = "paid")
}

