% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exhibit.ata}
\alias{exhibit.ata}
\title{Return a triangle for age to age development factors}
\usage{
\method{exhibit}{ata}(ata, selection = NULL, tail_column = FALSE)
}
\arguments{
\item{ata}{object of class ata generated from \code{ChainLadder} package}

\item{selection}{optional selected development factors}

\item{tail_column}{optional column for development factor of
most mature age to ultimate}
}
\description{
Return a triangle for age to age development factors
}
\examples{
library(ChainLadder)
tri <- as.triangle(recovery_ldf, origin = "origin",
                   dev = "dev", value = "paid_loss_only")
dev_tri <- ata(tri)
exhibit(dev_tri, selection = c(1.9, 1.2, 1.13, 1.075, NA))

# with tail factor selected
exhibit(dev_tri, selection = c(1.9, 1.2, 1.13, 1.075, 1.1))
}
\keyword{internal}

