% Generated by roxygen2 (4.0.1): do not edit by hand
\name{exhibit}
\alias{exhibit}
\title{Generic similar to summary}
\usage{
exhibit(data, ...)
}
\arguments{
\item{data}{objext to turn into exhibit}

\item{...}{additional arguments}
}
\description{
Exhibit defines certain ways to arrange values in
the `reserve` package for presentation in a report.
It is similar to summary, but tailored to fit nicely
on a single sheet of paper...
}
\examples{
tri <- as.triangle(recovery_ldf, origin = "origin", dev = "dev", value = "paid_loss_only")
dev_tri <- ata(tri)
exhibit(dev_tri, selection = c(1.9, 1.2, 1.13, 1.075))
}

