% Generated by roxygen2 (4.0.1): do not edit by hand
\name{exhibit.triangle}
\alias{exhibit.triangle}
\title{Returns an age to age development triangle}
\usage{
\method{exhibit}{triangle}(tri)
}
\arguments{
\item{ata}{object of class triangle generated from \code{ChainLadder} package}
}
\description{
Returns an age to age development triangle
}
\keyword{internal}

