% Generated by roxygen2 (4.0.1): do not edit by hand
\name{exhibit.triangle}
\alias{exhibit.triangle}
\title{Returns a cleaner development triangle for use in reports}
\usage{
\method{exhibit}{triangle}(tri)
}
\arguments{
\item{tri}{object of class triangle generated from \code{ChainLadder} package}
}
\description{
Returns a cleaner development triangle for use in reports
}
\keyword{internal}

