% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/incremental.R
\name{incremental}
\alias{incremental}
\title{Transforms cumulative dollar value in a loss_df object into incremental}
\usage{
incremental(ldf, dollar, include_all = FALSE)
}
\arguments{
\item{ldf}{object of class loss_df}

\item{dollar}{column to transform from cumulative to incremental. Can be of any of
the allowable `dollar` types for class \code{loss_df} or any of the columns in the `dollar`
type for class \code{loss_df}.}

\item{include_all}{logical. default = FALSE.  Whether or not to calculate incremental
losses for development years where data from prior development year is missing.}
}
\description{
Transforms cumulative dollar value in a loss_df object into incremental
}
\details{
\code{incremental} turns cumulative loss values into incemental amounts.
Claims that are not reported until after the first development period after their
origin are given an incremental payment in the first at the time they are first reported.
}
\examples{
incremental(test_df, dollar = "paid")

# nexy two compare the effect of `include_all` argument
incremental(test_df_incomplete, dollar = "paid")
incremental(test_df_incomplete, dollar = "paid", include_all = TRUE)
}

