% Generated by roxygen2 (4.0.2): do not edit by hand
\name{incurred}
\alias{incurred}
\title{returns the total gross incurred amounts per claim}
\usage{
incurred(ldf)
}
\arguments{
\item{ldf}{a data frame of class loss_df}
}
\description{
returns the total gross incurred amounts per claim
}
\examples{
paid(recovery_ldf)
}

