% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sum_dollar.R
\name{incurred}
\alias{incurred}
\title{returns a vector for total gross incuured}
\usage{
incurred(ldf)
}
\arguments{
\item{ldf}{a data frame of class loss_df}
}
\description{
returns a vector for total gross incuured
}
\examples{
incurred(recovery_ldf)
}

