% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loss_df}
\alias{loss_df}
\title{loss_df S3 class constructor}
\usage{
loss_df(ldf, origin, dev, id = NULL, evaluation_date = NULL, paid = NULL,
  incurred = NULL, paid_recovery = NULL, incurred_recovery = NULL,
  desc = NULL)
}
\arguments{
\item{origin}{time period in which the claim originated}

\item{dev}{development stage of claim at relevant 'evaluation_date'}

\item{id}{optional identification key of claim}

\item{evaluation_date}{optional column naming the evaluation dates}

\item{paid}{vector listing names or numbers for all paid loss columns}

\item{incurred}{vector listing names or numbers for all incurred loss columns}

\item{paid_recovery}{vector listing names of numbers for all paid recovery columns}

\item{incurred_recovery}{vector listing names or numbers for all incurred recovery columns}

\item{desc}{vector of all additional columns relevant to the analysis. Common examples are
  'open claims', and 'reported claims' (when claims are evaluated on an occurence basis)}

\item{df}{a data frame containing the raw data of losses by claim}
}
\description{
\code{loss_df} is the primary S3 data class for the \code{lossdb}
package.  A \code{loss_df} is a data frame for insurance loss data.  The data frame
is structured on a claim or occurence basis (i.e. each row represents 1 claim
or 1 claim occurence) evaluated at a specific time, and each column contains
certain variables attributable to the claim at that evaluation time.
}
\details{
\code{loss_df} is designed to assign descriptive and consistent names
to loss data to allow for easier data analysis. The relevant
columns can be assigned to the above arguments using a character vector of the column names
or a numeric vector of the column position.
}
\examples{
# create loss_df object from `occurences` dataset
mydf <- loss_df(occurences,  id = "claim_number",
                             origin = "origin",
                             dev = "dev",
                             paid = c("paid_loss_only", "paid_expense"),
                             incurred = c("incurred_loss_only", "incurred_expense"),
                             paid_recovery = c("sal_sub", "paid_excess250"),
                             incurred_recovery = c("incurred_excess250", "sal_sub_incurred"),
                             desc = "claim_cts"
                 )
}

