\name{losses}
\alias{losses}
\title{example workers' compensation losses.}
\format{A data frame with 11,104 observations on 47 variables.}
\usage{
losses
}
\description{
This claims data is in a 'raw' state; it is constructed
of 47 columns, most of which are not relevant to a reserve analysis.  The
\code{losses} data set is meant to resemble the combination of 5 typical data sets
an actuary would receive from a small workers' compensation insurance provider.  The \code{losses}
claims data is evaluated at 2009-06-30, 2010-06-30, 2011-06-30, 2012-06-30, and
2013-06-30.
}

