% Generated by roxygen2 (4.0.1): do not edit by hand
\name{paid}
\alias{paid}
\title{returns the total gross paid amounts per claim}
\usage{
paid(ldf)
}
\arguments{
\item{ldf}{a data frame of class loss_df}
}
\description{
returns the total gross paid amounts per claim
}
\examples{
paid(recovery_ldf)
}

