% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.loss_df}
\alias{plot.loss_df}
\title{plot method for \code{loss_df}}
\usage{
\method{plot}{loss_df}(df, evaluation_date = NULL)
}
\arguments{
\item{df}{loss_df S3 object}
}
\description{
barchart of loss amounts by origin year
}
\examples{
# plot paid and case at most recent `evaluation_date`
plot(losses_loss_df)

# plot paid and case at selected `evaluation_date`
plot(losses_loss_df, evaluation_date = "2012-06-30")
## need to fix so other evaluation dates besides latest work
}
\keyword{internal}

