% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sum_dollar}
\alias{sum_dollar}
\title{returns the sum of the selected loss_df type}
\usage{
sum_dollar(ldf, dollar)
}
\arguments{
\item{ldf}{a data frame of class loss_df}

\item{dollar}{the dollar type to be summed up. `dollar` options are `paid`,
`incurred`, `paid_recovery`, and `incurred_recovery`.}
}
\description{
returns the sum of the selected loss_df type
}
\examples{
sum_dollar(ldf = ldf_data, dollar = "paid")
}

