% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.loss_df}
\alias{summary.loss_df}
\title{returns the grouped sum of loss_df columns by origin period}
\usage{
\method{summary}{loss_df}(ldf, values = NULL, evaluation_date = NULL)
}
\arguments{
\item{ldf}{loss_df S3 object}

\item{values}{optional - select specific values to summarize}

\item{evaluation_date}{optional evaluation date}
}
\description{
summary.loss_df can provides snapshot for any columns
or `type` attributes in a `loss_df` object.
}
\examples{
# without specificied `evaluation_date`
summary(recovery_ldf)

# with specified `evaluation_date`
summary(recovery_ldf, evaluation_date = "2012-06-30")

# with specified `values`
summary(recovery_ldf, values = c("paid_excess250", "sal_sub", "paid"))
}

