% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.loss_df}
\alias{summary.loss_df}
\title{returns the grouped sum of loss_df columns by origin period}
\usage{
\method{summary}{loss_df}(ldf, values = NULL, calendar = NULL)
}
\arguments{
\item{ldf}{loss_df S3 object}

\item{values}{optional - select specific values to summarize}

\item{calendar}{optional calendar period (i.e. calendar = origin + dev)}

\item{by}{type attribute to summarize by.  Allowable values are `origin`, `dev`,
and `id`}
}
\description{
summary.loss_df can provides snapshot for any columns
or `type` attributes in a `loss_df` object.
}
\examples{
# without specificied `calendar`
summary(ldf_data)

# with specified `calendar`
summary(ldf_data, calendar = 2012)

# with specified `values`
summary(ldf_data, values = c("paid_excess250", "sal_sub", "paid"))
}

