% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.loss_df}
\alias{summary.loss_df}
\title{returns the grouped sum of loss_df columns by origin period}
\usage{
\method{summary}{loss_df}(ldf, detail = NULL, calendar = NULL)
}
\arguments{
\item{ldf}{object of class \code{loss_df}}

\item{detail}{optional - select specific detail names or columns names to include}

\item{calendar}{optional - calendar period (i.e. calendar = origin + dev)}
}
\description{
returns the grouped sum of loss_df columns by origin period
}
\examples{
# without specificied `calendar`
summary(ldf_data)

# with specified `calendar`
summary(ldf_data, calendar = 2012)

# with specified `detail`
summary(ldf_data, detail = c("paid_excess250", "sal_sub", "paid"))
}

