% Generated by roxygen2 (4.0.0): do not edit by hand
\name{summary.loss_df}
\alias{summary.loss_df}
\title{get the sum of loss_df data by origin period}
\usage{
\method{summary}{loss_df}(df, evaluation_date = NULL)
}
\arguments{
\item{loss_df}{loss_df S3 object}
}
\description{
get the sum of loss_df data by origin period
}
\examples{
# without specificied `evaluation_date`
summary(losses_loss_df)

# with specified `evaluation_date`
summary(losses_loss_df, evaluation_date = "2012-06-30")
}
\keyword{internal}

